refactor(enum): split enumeration.rs tests and entries-shape under the cap
